+++ cluster_tools/cluster_tools/schedulers/kube.py
@@ -81,6 +81,15 @@ class KubernetesExecutor(ClusterExecutor):
             return job_id
         return cls.get_jobid_with_index(job_id, job_index)
 
+    def inner_handle_kill(self, *args, **kwargs):
+        job_ids = ",".join(str(job_id) for job_id in self.jobs.keys())
+
+        print(
+            "Couldn't automatically cancel all Kubernetes jobs. The following jobs are still running on the cluster:\n{}".format(
+                job_ids
+            )
+        )
+
     def ensure_kubernetes_namespace(self):
         kubernetes_client = KubernetesClient()
         try: